pyproject.toml
[project]
name = "log-analyses-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "altair==4.2.2",
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.15",
    "langchain-text-splitters>=1.1.1",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "openai>=2.24.0",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.1",
    "rank-bm25>=0.2.2",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.2.3",
    "streamlit>=1.19.0",
    "torch>=2.10.0",
    "tqdm>=4.67.3",
    "transformers>=5.2.0",
]
